\name{croton}
\alias{croton}
\docType{data}
\title{
Replicated Point Pattern of Croton 
}
\description{
A list with a) a list of 16 point patterns  (with the \link[spatstat]{ppp} format of \code{spatstat}) of \emph{Croton wagneri} in Soutern Ecuador; b) a factor  with different elevations ("high", "slow")  and c) a factor with different topographical conditions ("steep" or "flat" slope) for each point pattern. Each point pattern is actually the result of a random thining (50 percent) of the original  pattern analyzed by  Ramon et al. (in revision).
}
\usage{data("croton")}

\references{
Ramon, P., De la Cruz, M., Chacon-Labella, J. & Escudero, A. (in revision). A new two-way ANOVA-like method for analyzing replicated point patterns in ecology.
}
\examples{
data(croton)
}
\keyword{datasets}
